fixed ambiguity about python script file locations
First, the readme instructions for scp'ing the python scripts are a little bit vague on path, specifically

a) the systemd unit files need the scripts to be in:

 /home/pi/erras/erras_rfid_reader.py
 /home/pi/erras/erras_members.py

But the way the readme was written it can be read as 

 /home/pi/erras/erras/erras_rfid_reader.py
 /home/pi/erras/erras/erras_members.py

I've edited my copy of the readme to deal with that.

Second, you need to disable serial login and enable serial port using raspi-config.  This is a standard rpi thing and really outside the scope of Erras but it seems like a good idea to add a bit to the readme about that.